+++ tests/test_bundle.py
@@ -18,7 +18,7 @@ import os
 import unittest
 import subprocess
 
-from sugar3.bundle import bundle_from_dir, bundle_from_archive
+from sugar3.bundle.helpers import bundle_from_dir, bundle_from_archive
 from sugar3.bundle.activitybundle import ActivityBundle
 from sugar3.bundle.contentbundle import ContentBundle
